+++ tiger-test-lib/src/test/java/de/gematik/test/tiger/lib/rbel/RbelMessageValidatorTest.java
@@ -82,8 +82,8 @@ class RbelMessageValidatorTest {
     when(tigerProxy.getRbelMessages()).thenReturn(validatableMessagesMock);
     rbelMessageValidator = new RbelMessageValidator(mock(TigerTestEnvMgr.class), tigerProxy);
 
-    rbelMessageValidator.currentRequest = null;
-    rbelMessageValidator.currentResponse = null;
+    RbelMessageValidator.currentRequest = null;
+    RbelMessageValidator.currentResponse = null;
     LocalProxyRbelMessageListener.getInstance().clearValidatableRbelMessages();
   }
 
@@ -330,7 +330,7 @@ class RbelMessageValidatorTest {
     addTwoRequestsToTigerTestHooks(validatableMessagesMock);
     RbelMessageValidator validator = rbelMessageValidator;
     validator.filterRequestsAndStoreInContext(RequestParameter.builder().path(".*").build());
-    RbelElement request = validator.currentRequest;
+    RbelElement request = RbelMessageValidator.currentRequest;
     assertTrue(validator.doesHostMatch(request, "localhost:8080"));
   }
 
@@ -351,7 +351,7 @@ class RbelMessageValidatorTest {
 
     validator.filterRequestsAndStoreInContext(
         RequestParameter.builder().path(".*").startFromLastRequest(true).build());
-    RbelElement request = validator.currentRequest;
+    RbelElement request = RbelMessageValidator.currentRequest;
 
     assertTrue(validator.doesHostMatch(request, "eitzen.at:80"));
   }
@@ -361,7 +361,7 @@ class RbelMessageValidatorTest {
     addTwoRequestsToTigerTestHooks(validatableMessagesMock);
     RbelMessageValidator validator = rbelMessageValidator;
     validator.findLastRequest();
-    RbelElement request = validator.currentRequest;
+    RbelElement request = RbelMessageValidator.currentRequest;
     assertTrue(validator.doesHostMatch(request, "eitzen.at:80"));
   }
 
@@ -375,7 +375,7 @@ class RbelMessageValidatorTest {
             .rbelPath("$.header.User-Agent")
             .value("mypersonalagent")
             .build());
-    assertTrue(validator.doesHostMatch(validator.currentRequest, "eitzen.at:80"));
+    assertTrue(validator.doesHostMatch(RbelMessageValidator.currentRequest, "eitzen.at:80"));
   }
 
   @Test
@@ -404,7 +404,7 @@ class RbelMessageValidatorTest {
             .rbelPath("$.header.User-Agent")
             .value("mypersonal.*")
             .build());
-    assertTrue(validator.doesHostMatch(validator.currentRequest, "eitzen.at:80"));
+    assertTrue(validator.doesHostMatch(RbelMessageValidator.currentRequest, "eitzen.at:80"));
   }
 
   @Test
@@ -413,7 +413,7 @@ class RbelMessageValidatorTest {
     RbelMessageValidator validator = rbelMessageValidator;
     validator.filterRequestsAndStoreInContext(
         RequestParameter.builder().path(".*").rbelPath("$.header.User-Agent").build());
-    assertTrue(validator.doesHostMatch(validator.currentRequest, "localhost:8080"));
+    assertTrue(validator.doesHostMatch(RbelMessageValidator.currentRequest, "localhost:8080"));
   }
 
   @Test
@@ -422,7 +422,7 @@ class RbelMessageValidatorTest {
     RbelMessageValidator validator = rbelMessageValidator;
     validator.filterRequestsAndStoreInContext(
         RequestParameter.builder().path(".*").rbelPath("$.header.Eitzen-Specific-header").build());
-    assertTrue(validator.doesHostMatch(validator.currentRequest, "eitzen.at:80"));
+    assertTrue(validator.doesHostMatch(RbelMessageValidator.currentRequest, "eitzen.at:80"));
   }
 
   @Test
@@ -443,10 +443,9 @@ class RbelMessageValidatorTest {
             .rbelPath("$.header.User-Agent")
             .value("mypersonal.*")
             .build());
-    assertTrue(validator.doesHostMatch(validator.currentRequest, "eitzen.at:80"));
+    assertTrue(validator.doesHostMatch(RbelMessageValidator.currentRequest, "eitzen.at:80"));
     assertThat(
-            validator
-                .currentResponse
+            RbelMessageValidator.currentResponse
                 .getFacet(RbelHttpResponseFacet.class)
                 .get()
                 .getResponseCode()
@@ -535,7 +534,7 @@ class RbelMessageValidatorTest {
     final RbelElement convertedMessage =
         rbelConverter.parseMessage(
             challengeMessage.getBytes(), null, null, Optional.of(ZonedDateTime.now()));
-    validator.currentResponse = convertedMessage;
+    RbelMessageValidator.currentResponse = convertedMessage;
     validatableMessagesMock.add(convertedMessage);
 
     // validate
@@ -617,7 +616,7 @@ class RbelMessageValidatorTest {
     TigerGlobalConfiguration.putValue("value.from.config1", "hello");
     TigerGlobalConfiguration.putValue("value.from.config2", "world");
 
-    rbelMessageValidator.currentRequest =
+    RbelMessageValidator.currentRequest =
         RbelLogger.build()
             .getRbelConverter()
             .parseMessage(responseToCheck.getBytes(), null, null, Optional.of(ZonedDateTime.now()));
@@ -637,7 +636,7 @@ class RbelMessageValidatorTest {
     final RbelElement convertedMessage =
         rbelConverter.parseMessage(
             challengeMessage.getBytes(), null, null, Optional.of(ZonedDateTime.now()));
-    validator.currentRequest = convertedMessage;
+    RbelMessageValidator.currentRequest = convertedMessage;
     validatableMessagesMock.add(convertedMessage);
 
     validator.assertAttributeOfCurrentRequestMatches("$.body.foo", "blala", false);
@@ -652,7 +651,7 @@ class RbelMessageValidatorTest {
     final RbelElement convertedMessage =
         rbelConverter.parseMessage(
             challengeMessage.getBytes(), null, null, Optional.of(ZonedDateTime.now()));
-    validator.currentRequest = convertedMessage;
+    RbelMessageValidator.currentRequest = convertedMessage;
     validatableMessagesMock.add(convertedMessage);
     assertThatThrownBy(
             () -> validator.assertAttributeOfCurrentRequestMatches("$.body.foo", "blabla", true))
@@ -729,7 +728,7 @@ class RbelMessageValidatorTest {
     await().until(() -> searchThread.getState() == Thread.State.WAITING);
     waitForParsing.complete(null);
 
-    assertThat(rbelMessageValidator.getCurrentResponse())
+    assertThat(RbelMessageValidator.getCurrentResponse())
         .extractChildWithPath("$.responseCode")
         .hasStringContentEqualTo("200");
   }
@@ -745,8 +744,8 @@ class RbelMessageValidatorTest {
     rbelMessageValidator.filterRequestsAndStoreInContext(
         RequestParameter.builder().rbelPath("$.path").value("/VAU").build());
 
-    log.info("current request: {} ", http(rbelMessageValidator.currentRequest));
-    log.info("current response: {} ", http(rbelMessageValidator.currentResponse));
+    log.info("current request: {} ", http(RbelMessageValidator.currentRequest));
+    log.info("current response: {} ", http(RbelMessageValidator.currentResponse));
 
     // next request, which comes immediately after the first one, no response in between
     rbelMessageValidator.filterRequestsAndStoreInContext(
@@ -756,10 +755,10 @@ class RbelMessageValidatorTest {
             .build()
             .resolvePlaceholders());
 
-    log.info("current request: {} ", http(rbelMessageValidator.currentRequest));
-    log.info("current response: {} ", http(rbelMessageValidator.currentResponse));
+    log.info("current request: {} ", http(RbelMessageValidator.currentRequest));
+    log.info("current response: {} ", http(RbelMessageValidator.currentResponse));
 
-    assertThat(rbelMessageValidator.getCurrentRequest())
+    assertThat(RbelMessageValidator.getCurrentRequest())
         .extractChildWithPath("$.path")
         .hasStringContentEqualTo("/1716066754997");
   }
@@ -788,7 +787,7 @@ class RbelMessageValidatorTest {
     final RbelElement convertedMessage =
         rbelConverter.parseMessage(
             challengeMessage.getBytes(), null, null, Optional.of(ZonedDateTime.now()));
-    validator.currentResponse = convertedMessage;
+    RbelMessageValidator.currentResponse = convertedMessage;
     validatableMessagesMock.add(convertedMessage);
 
     return validator;