+++ ReleaseNotes.md
@@ -4,6 +4,8 @@
 
 ## Bugfixes
 
+* TGR-1438: Tiger-Test-Lib: make currentResponse and currentRequest in RbelMessageValidator static so that they are
+  shared by multiple instances of RbelMessageValidator.
 * TGR-1439: Tiger-Test-Lib: made usage of LocalProxyRbelMessageListener null safe also when the local tiger proxy is not
   active.
 
@@ -81,7 +83,6 @@ For more details on how to specify a json schema refer to the external
 resource https://jsoneditoronline.org/indepth/validate/json-schema-validator/ .
 In https://json-schema.org/implementations#validators-web-(online) you can find a list of online validators which you
 can use to prepare the schemas. In Tiger we use the **2020-12** version of the JSON Schema specification.
-
 * TGR-1404: Tiger-Proxy: in direct reverse proxy mode the proxy opens immediately a connection to the remote target as
   soon as an incoming connection is established.